pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "AbstractSearch"
version = "0.1.0"
description = "Package to search CAR-related abstracts"
dependencies = [
    "rank_bm25",
    "sentence-transformers",
    "faiss-cpu",
    "nltk",
    "pandas",
    "numpy"
]

[tool.setuptools]
packages = ["AbstractSearch"]